backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
cloudinary==1.44.1
coverage==7.13.4
dj-database-url==3.1.2
Django==6.0.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-nested-admin==4.1.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.29.0
gunicorn==25.3.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pluggy==1.6.0
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-dotenv==1.2.2
python-monkey-business==1.1.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
